requirements.txt
fastapi
uvicorn
scikit-learn
pandas
pytest
httpx
google-cloud-storage
google-cloud-bigquery
db-dtypes
fsspec
gcsfs
matplotlib